Move the new layer form into a new component
