+++ src/components/sidebar-entities/layer-sidebar/index.js
@@ -1,27 +1,12 @@
 import { connect } from 'react-redux';
-import { injectIntl } from 'react-intl';
 import { createStructuredSelector } from 'reselect';
 
-import {
-  currentEntitySelector,
-  layerFormSelector,
-} from 'store/selectors/base.selectors';
-import { isValidLayerForm } from 'store/selectors/layer.selectors';
-import { updateLayer, addLayer } from 'store/actions/layer.actions';
+import { currentEntitySelector } from 'store/selectors/base.selectors';
 
 import LayerSidebar from './layer-sidebar';
 
 const mapStateToProps = createStructuredSelector({
-  entity: currentEntitySelector,
-  layerForm: layerFormSelector,
-  isValid: isValidLayerForm,
+  layer: currentEntitySelector,
 });
 
-const mapDispatchToProps = (dispatch, props) => ({
-  updateLayer: (key, value) => dispatch(updateLayer(key, value)),
-  addLayer: () => dispatch(addLayer(props.intl)),
-});
-
-export default injectIntl(
-  connect(mapStateToProps, mapDispatchToProps)(LayerSidebar),
-);
+export default connect(mapStateToProps)(LayerSidebar);
