+++ src/components/sidebar-entities/layer-sidebar/new-layer/style.scss
@@ -1,7 +1,7 @@
-.LayerSidebar {
+.NewLayer {
   margin: 1rem;
 }
 
-.LayerSidebar-Create {
+.NewLayer-Create {
   margin: 1rem 0 0;
 }
